requirements.txt
# --- Core Data Processing & Physics ---
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0

# --- Database Infrastructure ---
# Used for the "Virtual Memory" of the Digital Twin
mysql-connector-python>=8.2.0

# --- Deep Learning (RAMNet) ---
# Supports the 675-parameter model architecture
torch>=2.1.0
scikit-learn>=1.3.0

# --- Visualization & Progress Monitoring ---
# Essential for the 7,000-cycle timeline and TQDM bars
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0